+++ fiddles/vue/fiddle-0016-AgGridContextMenuIconJs/src/main.js
@@ -1,0 +1,19 @@
+import { createApp } from 'vue'
+
+import App from './App.vue'
+import router from './router'
+
+import './assets/main.css'
+import 'bootswatch/dist/sketchy/bootstrap.css'
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { faEllipsisV } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+library.add(faEllipsisV)
+
+const app = createApp(App)
+
+app.component('font-awesome-icon', FontAwesomeIcon)
+
+app.use(router)
+
+app.mount('#app')