+++ fiddles/vue/fiddle-0016-AgGridContextMenuIconJs/src/views/DocsView.vue
@@ -1,0 +1,18 @@
+<script setup>
+import DocList from '../components/DocList.vue'
+</script>
+
+<template>
+  <div class="content-container">
+    <DocList />
+  </div>
+</template>
+
+<style scoped>
+.content-container {
+  margin-top: 1rem;
+  flex-grow: 1;
+  overflow-y: auto;
+  padding: 0.5rem;
+}
+</style>