+++ CHANGELOG.md
@@ -2555,7 +2555,16 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 
 ### 2050-python-qt-to-jpeg-file-converter ~ [Issue 2050](https://github.com/bradyhouse/house/issues/2050)
 
-* **Python**
+* **python**
 
   1. Added [fiddles/python/fiddle-0015-QuicktimeSplitter](fiddles/python/fiddle-0015-QuicktimeSplitter)
+
+
+### 1997-vue-aggrid-row-context-icon ~ [Issue 1997](https://github.com/bradyhouse/house/issues/1997)
+
+* **vue**
+
+  1. Added [fiddles/vue/fiddle-0016-AgGridContextMenuIconJs](fiddles/vue/fiddle-0016-AgGridContextMenuIconJs)
+
+